system_v2/rpi2/requirements.txt
pymodbus>=3.5.0
pyiec61850>=1.5.2
asyncio
pyyaml
